ProfileDataのtoModelメソッドを移動する
